# client/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5001"
    }
  ],
  "paths": {
...
    "/Game/NewGame": {
      "post": {
        "tags": [
          "Game"
        ],
        "operationId": "Game_NewGame",
        "requestBody": {
          "x-name": "prize",
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/Game/NewGameFromMonday": {
      "post": {
        "tags": [
          "Game"
        ],
        "operationId": "Game_NewGameFromMonday",
        "requestBody": {
          "x-name": "prize",
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/Game/winningNumbers": {
      "post": {
        "tags": [
          "Game"
        ],
        "operationId": "Game_SetWinningNumbers",
        "requestBody": {
          "x-name": "data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WinningNumbersRequestDTO"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WinningNumbersResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GameResponseDTO": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "$ref": "#/components/schemas/GameStatus"
          }
        }
      },
      "GameStatus": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Active",
          "Inactive"
        ],
        "enum": [
          "Active",
          "Inactive"
        ]
      },
      "WinningNumbersResponseDTO": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "gameid": {
            "type": "string",
            "format": "guid"
          },
          "winningnumbers": {
            "type": "array",
            "maxItems": 3,
            "minItems": 3,
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "status": {
            "$ref": "#/components/schemas/GameStatus"
          }
        }
      },
      "WinningNumbersRequestDTO": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "gameId",
          "winningNumbers"
        ],
        "properties": {
          "gameId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          },
          "winningNumbers": {
            "type": "array",
            "maxItems": 3,
            "minItems": 3,
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
...
    }
  }
}